turning prices to integers

+++ database/migrations/2024_11_06_140457_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('order_id')->nullable()->constrained()->cascadeOnDelete()->cascadeOnUpdate();
             $table->foreignId('product_id')->nullable()->constrained()->cascadeOnDelete()->cascadeOnUpdate();
-            $table->double('total')->default(0);
+            $table->integer('total')->default(0);
             $table->integer('qty')->default(1);
             $table->json('attributes')->nullable();
             $table->timestamps();
